requirements.txt
feedparser>=6.0
discord-webhook>=1.3
google-genai>=1.0
python-dotenv>=1.0
schedule>=1.2
requests>=2.31
